+++ CHANGELOG.md
@@ -29,7 +29,7 @@
   (PR[#3634](https://github.com/scality/metalk8s/pull/3634))
 
 - Bump containerd to [1.6.4](https://github.com/containerd/containerd/releases/tag/v1.6.4)
-  The pause image has been bump to 3.6
+  The pause image has been bump to 3.7
   (PR[#3778](https://github.com/scality/metalk8s/pull/3778))
 
 - Bump Calico version to [3.23.1](https://github.com/projectcalico/calico/releases/tag/v3.23.1)